requirements.txt
Django==5.2.6
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.7
stripe==6.6.0
django-allauth==65.13.0
python-dotenv==1.2.1
cryptography==46.0.3
pyjwt==2.10.1
brotli==1.2.0
channels==4.0.0
daphne==4.0.0
channels-redis==4.1.0
openai==1.54.0
redis==5.0.1
openpyxl==3.1.2
xlsxwriter==3.1.9